vercel_fix

+++ packages/foundry/contracts/YourContract.sol
@@ -11,33 +11,32 @@ import "forge-std/console.sol";
  * to create a transparent accountability layer.
  */
 contract YourContract {
-
     // --- Data Structures ---
 
     /**
      * @notice The complete data structure for a single case.
      */
     struct CaseReport {
-        uint id;                 // The unique case ID
-        address citizenReporter;  // The invisible wallet of the filer
-        address assignedPolice;   // The police wallet that acknowledged it
+        uint256 id; // The unique case ID
+        address citizenReporter; // The invisible wallet of the filer
+        address assignedPolice; // The police wallet that acknowledged it
         address verifyingJudicial; // The judicial wallet that stamped it
-        uint timestampFiled;      // The block.timestamp when filed
-        uint timestampLastUpdate; // The timestamp of the last update
-        string caseDetails;       // The encrypted/hashed details of the report
-        string location;          // Public location (e.g., "District 7")
-        CaseStatus status;        // The current status from our enum
-        string[] publicUpdates;   // A log of all updates
+        uint256 timestampFiled; // The block.timestamp when filed
+        uint256 timestampLastUpdate; // The timestamp of the last update
+        string caseDetails; // The encrypted/hashed details of the report
+        string location; // Public location (e.g., "District 7")
+        CaseStatus status; // The current status from our enum
+        string[] publicUpdates; // A log of all updates
     }
 
     /**
      * @notice The distinct lifecycle stages of a case.
      */
     enum CaseStatus {
-        Filed,            // 0: Just filed by a citizen
-        Acknowledged,     // 1: Acknowledged by a Police wallet
-        ClosedByPolice,   // 2: Resolved by Police, pending verification
-        VerifiedClosed    // 3: Final verification by Judiciary
+        Filed, // 0: Just filed by a citizen
+        Acknowledged, // 1: Acknowledged by a Police wallet
+        ClosedByPolice, // 2: Resolved by Police, pending verification
+        VerifiedClosed // 3: Final verification by Judiciary
     }
 
     // --- State Variables ---
@@ -48,8 +47,8 @@ contract YourContract {
     mapping(address => bool) public isJudicial;
 
     // The "Database"
-    uint public caseCounter;
-    mapping(uint => CaseReport) public cases;
+    uint256 public caseCounter;
+    mapping(uint256 => CaseReport) public cases;
 
     // --- Event ---
 
@@ -57,7 +56,7 @@ contract YourContract {
      * @notice Emitted every time a case changes state.
      * The frontend dashboard will listen for this.
      */
-    event CaseUpdated(uint indexed caseId, CaseStatus newStatus, string updateMessage);
+    event CaseUpdated(uint256 indexed caseId, CaseStatus newStatus, string updateMessage);
 
     // --- Modifiers ---
 
@@ -126,7 +125,7 @@ contract YourContract {
      */
     function fileReport(string memory _details, string memory _location) public {
         caseCounter++;
-        uint _id = caseCounter;
+        uint256 _id = caseCounter;
 
         CaseReport storage _case = cases[_id];
         _case.id = _id;
@@ -146,7 +145,7 @@ contract YourContract {
     /**
      * @notice Police: Acknowledge a 'Filed' case.
      */
-    function acknowledgeCase(uint _caseId, string memory _updateMessage) public onlyPolice {
+    function acknowledgeCase(uint256 _caseId, string memory _updateMessage) public onlyPolice {
         CaseReport storage _case = cases[_caseId];
         require(_case.status == CaseStatus.Filed, "CASE: Not in filed state");
 
@@ -154,21 +153,21 @@ contract YourContract {
         _case.assignedPolice = msg.sender;
         _case.timestampLastUpdate = block.timestamp;
         _case.publicUpdates.push(_updateMessage);
-        
+
         emit CaseUpdated(_caseId, _case.status, _updateMessage);
     }
 
     /**
      * @notice Police: Add a public progress update to an 'Acknowledged' case.
      */
-    function addProgressUpdate(uint _caseId, string memory _updateMessage) public onlyPolice {
+    function addProgressUpdate(uint256 _caseId, string memory _updateMessage) public onlyPolice {
         CaseReport storage _case = cases[_caseId];
         // Can only update cases that are actively being investigated
         require(_case.status == CaseStatus.Acknowledged, "CASE: Not in 'Acknowledged' state");
 
         _case.timestampLastUpdate = block.timestamp;
         _case.publicUpdates.push(_updateMessage);
-        
+
         // We emit the same event, just with the new message
         emit CaseUpdated(_caseId, _case.status, _updateMessage);
     }
@@ -176,21 +175,21 @@ contract YourContract {
     /**
      * @notice Police: Close an 'Acknowledged' case.
      */
-    function closeCase(uint _caseId, string memory _finalReport) public onlyPolice {
+    function closeCase(uint256 _caseId, string memory _finalReport) public onlyPolice {
         CaseReport storage _case = cases[_caseId];
         require(_case.status == CaseStatus.Acknowledged, "CASE: Not acknowledged");
 
         _case.status = CaseStatus.ClosedByPolice;
         _case.timestampLastUpdate = block.timestamp;
         _case.publicUpdates.push(_finalReport);
-        
+
         emit CaseUpdated(_caseId, _case.status, _finalReport);
     }
 
     /**
      * @notice Judiciary: Verify a 'ClosedByPolice' case.
      */
-    function verifyClosure(uint _caseId, string memory _verificationMessage) public onlyJudicial {
+    function verifyClosure(uint256 _caseId, string memory _verificationMessage) public onlyJudicial {
         CaseReport storage _case = cases[_caseId];
         require(_case.status == CaseStatus.ClosedByPolice, "CASE: Not closed by police");
 
@@ -198,17 +197,17 @@ contract YourContract {
         _case.verifyingJudicial = msg.sender;
         _case.timestampLastUpdate = block.timestamp;
         _case.publicUpdates.push(_verificationMessage);
-        
+
         emit CaseUpdated(_caseId, _case.status, _verificationMessage);
     }
 
     // --- 4. Viewer Functions (Read-Only) ---
 
-    function getCaseDetails(uint _caseId) public view returns (CaseReport memory) {
+    function getCaseDetails(uint256 _caseId) public view returns (CaseReport memory) {
         return cases[_caseId];
     }
 
-    function getTotalCases() public view returns (uint) {
+    function getTotalCases() public view returns (uint256) {
         return caseCounter;
     }
 }
